requirements.txt
# LangChain Core
langchain
langchain-core

# Langchain Experimental
langchain_experimental

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
sentence-transformers

# Perplexity Integration
langchain_perplexity
langchain_community   ## community

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Streamlit
streamlit

# Datatype Validations
pydantic
pydantic[email]

# Visualizing langchain chains
grandalf

## Loaders
PyPDF
pymupdf
bs4   ## beautiful soup for webbase loader

## Youtube Transcript
youtube_transcript_api

## Built-in tools
duckduckgo-search 
ddgs